refactor: split action-widgets.tsx and markdown.tsx
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.

action-widgets.tsx: split into action-widgets-config.ts (pure/RPC config
helpers), more-dropdown.tsx (self-contained floating-UI component),
use-widget-bar-responsive.ts (3-tier responsive-collapse hook), and
use-widget-drag-reorder.ts (drag-to-reorder state machine hook). 687 -> 363
lines.

markdown.tsx: split into markdown-mermaid.tsx, markdown-codeblock.tsx, and
markdown-media.tsx — independent custom-renderer components only referenced
via the markdownComponents map. 620 -> 351 lines.

tsc --noEmit clean; full vitest suite 1904/1908 (one pre-existing unrelated
flaky timeout, confirmed passing in isolation both before and after).

<!-- agentmux:agent_id=agentx -->
