+++ frontend/app/element/markdown.tsx
@@ -1,16 +1,9 @@
 // Copyright 2025-2026, AgentMux Corp.
 // SPDX-License-Identifier: Apache-2.0
 
-import { CopyButton } from "@/app/element/copybutton";
-import { writeText as clipboardWriteText } from "@/util/clipboard";
 import { ErrorBoundary } from "@/app/element/errorboundary";
 import { createContentBlockPlugin } from "@/app/element/markdown-contentblock-plugin";
-import {
-    MarkdownContentBlockType,
-    resolveRemoteFile,
-    resolveSrcSet,
-    transformBlocks,
-} from "@/app/element/markdown-util";
+import { transformBlocks } from "@/app/element/markdown-util";
 import { ALIGN_CLASS_REGEX, rehypeAlignToClass } from "@/app/element/rehype-align-to-class";
 import remarkMermaidToTag from "@/app/element/remark-mermaid-to-tag";
 import { TableBlock } from "@/app/element/table-block";
@@ -33,24 +26,10 @@ import remarkParse from "remark-parse";
 import remarkRehype from "remark-rehype";
 import { openLink } from "../store/global";
 import { rehypeLinkify } from "./rehype-linkify";
-import { IconButton } from "./iconbutton";
+import { Code, CodeBlock } from "./markdown-codeblock";
+import { MarkdownImg, MarkdownSource, WaveBlock } from "./markdown-media";
+import { Mermaid, MermaidErrorFallback } from "./markdown-mermaid";
 import "./markdown.scss";
-
-let mermaidInitialized = false;
-let mermaidInstance: any = null;
-
-const initializeMermaid = async () => {
-    if (!mermaidInitialized) {
-        const mermaid = await import("mermaid");
-        mermaidInstance = mermaid.default;
-        mermaidInstance.initialize({
-            startOnLoad: false,
-            theme: "dark",
-            securityLevel: "strict",
-        });
-        mermaidInitialized = true;
-    }
-};
 
 const Link = ({
     setFocusedHeading,
@@ -81,254 +60,6 @@ const Heading = ({ props, hnum }: { props: JSX.HTMLAttributes<HTMLHeadingElement
         <div id={(props as any).id} class={clsx("heading", `is-${hnum}`)}>
             {(props as any).children}
         </div>
-    );
-};
-
-let mermaidRenderCount = 0;
-
-const Mermaid = ({ chart }: { chart: string }) => {
-    let ref!: HTMLDivElement;
-    const [svgContent, setSvgContent] = createSignal<string | null>(null);
-    const [error, setError] = createSignal<string | null>(null);
-
-    onMount(() => {
-        let cancelled = false;
-        const renderMermaid = async () => {
-            try {
-                setError(null);
-                setSvgContent(null);
-
-                await initializeMermaid();
-                if (cancelled || !mermaidInstance) {
-                    return;
-                }
-
-                // Normalize the chart text
-                const normalizedChart = chart
-                    .replace(/<br\s*\/?>/gi, "\n")
-                    .replace(/\r\n?/g, "\n")
-                    .replace(/\n+$/, "");
-
-                const id = `mermaid-${++mermaidRenderCount}`;
-                const { svg } = await mermaidInstance.render(id, normalizedChart);
-                if (!cancelled) {
-                    setSvgContent(svg);
-                }
-            } catch (err: any) {
-                console.error("Error rendering mermaid diagram:", err);
-                if (!cancelled) {
-                    setError(err.message || String(err));
-                }
-            }
-        };
-
-        renderMermaid();
-        onCleanup(() => {
-            cancelled = true;
-        });
-    });
-
-    return (
-        <Show
-            when={!error()}
-            fallback={
-                <div class="mermaid error">
-                    <div style={{ color: "var(--error-color, #f44)", "margin-bottom": "8px" }}>
-                        Failed to render diagram
-                    </div>
-                    <pre style={{ "white-space": "pre-wrap", opacity: 0.7, "font-size": "0.85em" }}>{chart}</pre>
-                </div>
-            }
-        >
-            <Show
-                when={svgContent()}
-                fallback={<div class="mermaid">Loading diagram...</div>}
-            >
-                <div class="mermaid" ref={ref} innerHTML={svgContent()} />
-            </Show>
-        </Show>
-    );
-};
-
-const MermaidErrorFallback = ({ error, chart }: { error?: Error; chart: string }) => (
-    <div class="mermaid error">
-        <div style={{ color: "var(--error-color, #f44)", "margin-bottom": "8px" }}>Failed to render diagram</div>
-        <pre style={{ "white-space": "pre-wrap", opacity: 0.7, "font-size": "0.85em" }}>{chart}</pre>
-    </div>
-);
-
-const Code = ({ className = "", children }: { className?: string; children: any }) => {
-    if (/\blanguage-mermaid\b/.test(className)) {
-        const text = Array.isArray(children) ? children.join("") : String(children ?? "");
-        return (
-            <ErrorBoundary fallback={<MermaidErrorFallback chart={text} />}>
-                <Mermaid chart={text} />
-            </ErrorBoundary>
-        );
-    }
-    return <code class={className}>{children}</code>;
-};
-
-type CodeBlockProps = {
-    children: any;
-    onClickExecute?: (cmd: string) => void;
-};
-
-const CodeBlock = ({ children, onClickExecute }: CodeBlockProps) => {
-    const getTextContent = (children: any): string => {
-        if (typeof children === "string") {
-            return children;
-        } else if (Array.isArray(children)) {
-            return children.map(getTextContent).join("");
-        } else if (children && children.props && children.props.children) {
-            return getTextContent(children.props.children);
-        }
-        return "";
-    };
-
-    const handleCopy = async (e: MouseEvent) => {
-        let textToCopy = getTextContent(children);
-        textToCopy = textToCopy.replace(/\n$/, "");
-        await clipboardWriteText(textToCopy);
-    };
-
-    const handleExecute = (e: MouseEvent) => {
-        let textToCopy = getTextContent(children);
-        textToCopy = textToCopy.replace(/\n$/, "");
-        if (onClickExecute) {
-            onClickExecute(textToCopy);
-        }
-    };
-
-    return (
-        <pre class="codeblock">
-            {children}
-            <div class="codeblock-actions">
-                <CopyButton onClick={handleCopy} title="Copy" />
-                {onClickExecute && (
-                    <IconButton
-                        decl={{
-                            elemtype: "iconbutton",
-                            icon: "regular@square-terminal",
-                            click: handleExecute,
-                        }}
-                    />
-                )}
-            </div>
-        </pre>
-    );
-};
-
-const MarkdownSource = ({
-    props,
-    resolveOpts,
-}: {
-    props: JSX.SourceHTMLAttributes<HTMLSourceElement> & {
-        srcSet?: string;
-        media?: string;
-    };
-    resolveOpts: MarkdownResolveOpts;
-}) => {
-    const [resolvedSrcSet, setResolvedSrcSet] = createSignal<string>((props as any).srcSet ?? "");
-    const [resolving, setResolving] = createSignal<boolean>(true);
-
-    onMount(() => {
-        const resolvePath = async () => {
-            const resolved = await resolveSrcSet((props as any).srcSet, resolveOpts);
-            setResolvedSrcSet(resolved);
-            setResolving(false);
-        };
-        resolvePath();
-    });
-
-    return (
-        <Show when={!resolving()}>
-            <source srcset={resolvedSrcSet()} media={(props as any).media} />
-        </Show>
-    );
-};
-
-interface WaveBlockProps {
-    blockkey: string;
-    blockmap: Map<string, MarkdownContentBlockType>;
-}
-
-const WaveBlock = (props: WaveBlockProps) => {
-    const { blockkey, blockmap } = props;
-    const block = blockmap.get(blockkey);
-    if (block == null) {
-        return null;
-    }
-    const sizeInKB = Math.round((block.content.length / 1024) * 10) / 10;
-    const displayName = block.id.replace(/^"|"$/g, "");
-    return (
-        <div class="waveblock">
-            <div class="wave-block-content">
-                <div class="wave-block-icon">
-                    <i class="fas fa-file-code"></i>
-                </div>
-                <div class="wave-block-info">
-                    <span class="wave-block-filename">{displayName}</span>
-                    <span class="wave-block-size">{sizeInKB} KB</span>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-const MarkdownImg = ({
-    props,
-    resolveOpts,
-}: {
-    props: JSX.ImgHTMLAttributes<HTMLImageElement>;
-    resolveOpts: MarkdownResolveOpts;
-}) => {
-    const src = (props as any).src as string;
-    const srcSet = (props as any).srcSet as string;
-
-    const [resolvedSrc, setResolvedSrc] = createSignal<string | null>(src);
-    const [resolvedSrcSet, setResolvedSrcSet] = createSignal<string | null>(srcSet ?? null);
-    const [resolvedStr, setResolvedStr] = createSignal<string | null>(null);
-    const [resolving, setResolving] = createSignal<boolean>(true);
-
-    onMount(() => {
-        if (src?.startsWith("data:image/")) {
-            setResolving(false);
-            setResolvedSrc(src);
-            setResolvedStr(null);
-            return;
-        }
-        if (resolveOpts == null) {
-            setResolving(false);
-            setResolvedSrc(null);
-            setResolvedStr(`[img:${src}]`);
-            return;
-        }
-
-        const resolveFn = async () => {
-            const [rSrc, rSrcSet] = await Promise.all([
-                resolveRemoteFile(src, resolveOpts),
-                resolveSrcSet(srcSet, resolveOpts),
-            ]);
-
-            setResolvedSrc(rSrc);
-            setResolvedSrcSet(rSrcSet);
-            setResolvedStr(null);
-            setResolving(false);
-        };
-        resolveFn();
-    });
-
-    return (
-        <Show when={!resolving()}>
-            <Show when={resolvedStr() != null} fallback={
-                <Show when={resolvedSrc() != null} fallback={<span>[img]</span>}>
-                    <img {...(props as any)} src={resolvedSrc()} srcset={resolvedSrcSet()} />
-                </Show>
-            }>
-                <span>{resolvedStr()}</span>
-            </Show>
-        </Show>
     );
 };
