+++ frontend/app/window/action-widgets-config.ts
@@ -1,0 +1,85 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Pure config helpers for the widget bar — pinned/more widget derivation and
+ * pin/unpin RPC persistence. Extracted from action-widgets.tsx.
+ *
+ * Settings keys:
+ *   widget:pinned   — ordered short-name array (e.g. ["agent","terminal","sysinfo"])
+ *   widget:icononly — icons only, no labels
+ */
+
+import { RpcApi } from "@/app/store/rpc-api";
+import { TabRpcClient } from "@/app/store/rpc-util";
+import { createBlock } from "@/store/global";
+import { fireAndForget } from "@/util/util";
+
+/**
+ * Return the effective pinned short-names (no "defwidget@" prefix), in order.
+ *
+ * Priority:
+ *  1. widget:pinned is set → authoritative
+ *  2. Not set → derive from display:pinned in widget config
+ */
+export function getPinnedKeys(
+    settings: Record<string, any>,
+    wmap: Record<string, WidgetConfigType>
+): string[] {
+    const pinned: string[] | undefined = settings["widget:pinned"];
+    if (pinned !== undefined) {
+        return pinned.filter((shortName) => wmap[`defwidget@${shortName}`] != null);
+    }
+    return Object.entries(wmap)
+        .filter(([, w]) => w["display:pinned"])
+        .sort(([, a], [, b]) => (a["display:order"] ?? 0) - (b["display:order"] ?? 0))
+        .map(([key]) => key.replace("defwidget@", ""));
+}
+
+export function getPinnedWidgets(
+    settings: Record<string, any>,
+    wmap: Record<string, WidgetConfigType>
+): { key: string; widget: WidgetConfigType }[] {
+    if (!wmap) return [];
+    return getPinnedKeys(settings, wmap)
+        .map((shortName) => {
+            const key = `defwidget@${shortName}`;
+            return { key, widget: wmap[key] };
+        })
+        .filter((e) => e.widget != null);
+}
+
+export function getMoreWidgets(
+    settings: Record<string, any>,
+    wmap: Record<string, WidgetConfigType>
+): { key: string; widget: WidgetConfigType }[] {
+    if (!wmap) return [];
+    const pinnedSet = new Set(getPinnedKeys(settings, wmap));
+    return Object.entries(wmap)
+        .filter(([key]) => !pinnedSet.has(key.replace("defwidget@", "")))
+        .sort(([, a], [, b]) => (a["display:order"] ?? 0) - (b["display:order"] ?? 0))
+        .map(([key, widget]) => ({ key, widget }));
+}
+
+// ── Widget actions ────────────────────────────────────────────────────────────
+
+export async function handleWidgetSelect(widget: WidgetConfigType) {
+    createBlock(widget.blockdef, widget.magnified);
+}
+
+export function pinWidget(shortName: string, settings: Record<string, any>, wmap: Record<string, WidgetConfigType>) {
+    fireAndForget(async () => {
+        const current = getPinnedKeys(settings, wmap);
+        if (current.includes(shortName)) return;
+        await RpcApi.SetConfigCommand(TabRpcClient, { "widget:pinned": [...current, shortName] } as any);
+    });
+}
+
+export function unpinWidget(shortName: string, settings: Record<string, any>, wmap: Record<string, WidgetConfigType>) {
+    fireAndForget(async () => {
+        const current = getPinnedKeys(settings, wmap);
+        await RpcApi.SetConfigCommand(TabRpcClient, {
+            "widget:pinned": current.filter((k) => k !== shortName),
+        } as any);
+    });
+}
